requirements.txt
# API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Azure OpenAI (LLM generation)
openai>=1.0.0
python-dotenv>=1.0.0

# Binary question-quality classifier (supervised learning; CS499)
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
